refactor: move config.py and database.py to app/ root per architecture v4
